README
In 2018 "fake news" and "promoted media" makes it very hard to filter out what is news we are supposed to see and what is news that is promoted by a news channel for money.  People are tired of fighting through promoted content and tracking cookie based shopping cart reminders. The answer ? A simple RSS Reader! RSS stands for Really Simple Syndication (or Rich Site Summary) and it was first stitched into the tapestry of the open web around the turn of the millennium. Its aim is straightforward: to make it easy to track updates to the content of a given website in a standardized format.